Add files via upload
I have raw data at certain intervals that are not all the same, and then the CSV is not setup to handle the same intervals, it just puts the values either in the start or not at all so to start this will be for interval of 15 min. The files labeled (raw data) is the actually reading on the graph and the file named (invalid) is the values that I get after downloading the csv. see code example.